requirements.txt
fastapi[standard]==0.115.0
PyJWT==2.9.0